# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Subnet Analytics Tournaments API",
    "description": "# Subnet Analytics Tournaments API\n\nPattern detection & feature generation evaluation platform for the Bittensor Subnet.\n\n## Tournament Scoring (Three Pillars)\n\nMiners are evaluated on **three pillars**:\n\n### 1. Features (10%)\n- Output schema validation for `features.parquet`\n- Generation performance relative to baseline\n\n### 2. Patterns (55%)\n- **Synthetic Recall (30%)**: % of ground_truth addresses detected\n- **Pattern Precision (25%)**: Anti-cheat via flow tracing validation\n\n### 3. Novelty (25%) + Performance (10%)\n- **Novelty Discovery (25%)**: New valid patterns beyond ground_truth\n- **Pattern Performance (10%)**: Detection speed relative to baseline\n\n## Flow Tracing Validation\n\nAll patterns (synthetic and novelty) are validated by verifying transaction flows exist in `transfers.parquet`. This prevents gaming through fake patterns.\n\n## API Versioning\n\nCurrent version: `v1`\n\nAll endpoints use `/api/v1/analytics/` prefix.\n\n## Authentication\n\nThis is a **read-only public API**. No authentication required.",
    "version": "1.0.0",
    "contact": {
      "name": "Subnet Team",
      "url": "https://github.com/subnet"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/api/v1/analytics/tournaments": {
      "get": {
        "tags": ["Tournaments"],
        "summary": "List Tournaments",
        "description": "Retrieve a paginated list of all analytics tournaments. Tournaments are ordered by epoch number in descending order (most recent first).\n\n**Use Cases:**\n- Display tournament history on dashboard\n- Load tournament selector dropdown\n- Monitor tournament status",
        "operationId": "list_tournaments",
        "parameters": [
          {
            "$ref": "#/components/parameters/SkipParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by tournament status",
            "schema": {
              "type": "string",
              "enum": ["pending", "running", "completed", "failed"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tournaments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/tournaments/{tournament_id}": {
      "get": {
        "tags": ["Tournaments"],
        "summary": "Get Tournament Details",
        "description": "Retrieve detailed information about a specific tournament including configuration, test networks, and baseline information.\n\n**Use Cases:**\n- Display tournament detail page\n- Load tournament configuration\n- Check tournament completion status",
        "operationId": "get_tournament",
        "parameters": [
          {
            "$ref": "#/components/parameters/TournamentIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tournament not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/tournaments/{tournament_id}/leaderboard": {
      "get": {
        "tags": ["Tournaments"],
        "summary": "Get Tournament Leaderboard",
        "description": "Retrieve the leaderboard for a specific tournament showing all miners ranked by final score.\n\nThe leaderboard includes:\n- Miner ranking and final scores\n- Component score breakdown (feature, synthetic, precision, novelty, performance)\n- Pattern detection statistics\n- Baseline comparison results\n\n**Use Cases:**\n- Display tournament leaderboard on dashboard\n- Show winner and top performers\n- Analyze score distribution",
        "operationId": "get_tournament_leaderboard",
        "parameters": [
          {
            "$ref": "#/components/parameters/TournamentIdParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament leaderboard with ranked miners",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tournament not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/tournaments/{tournament_id}/results": {
      "get": {
        "tags": ["Tournaments"],
        "summary": "Get Tournament Results",
        "description": "Retrieve all results for a specific tournament. Similar to leaderboard but includes additional aggregation data.\n\n**Use Cases:**\n- Export tournament results\n- Generate reports\n- Analyze detailed performance metrics",
        "operationId": "get_tournament_results",
        "parameters": [
          {
            "$ref": "#/components/parameters/TournamentIdParam"
          },
          {
            "$ref": "#/components/parameters/SkipParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentResultResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tournament not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/tournaments/{tournament_id}/submissions": {
      "get": {
        "tags": ["Tournaments"],
        "summary": "Get Tournament Submissions",
        "description": "Retrieve all miner submissions for a specific tournament.\n\n**Use Cases:**\n- View submission statuses\n- Monitor validation progress\n- Track miner participation",
        "operationId": "get_tournament_submissions",
        "parameters": [
          {
            "$ref": "#/components/parameters/TournamentIdParam"
          },
          {
            "$ref": "#/components/parameters/SkipParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament submissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tournament not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/tournaments/{tournament_id}/runs": {
      "get": {
        "tags": ["Tournaments"],
        "summary": "Get Evaluation Runs",
        "description": "Retrieve all evaluation runs for a specific tournament. Each run represents one miner's evaluation on one test network.\n\nRun data includes:\n- Schema validation results\n- Synthetic pattern recall metrics\n- Novelty pattern discovery metrics\n- Flow tracing validation results\n- Performance timing\n\n**Use Cases:**\n- Debug miner performance issues\n- Analyze per-network results\n- Track evaluation progress",
        "operationId": "get_tournament_runs",
        "parameters": [
          {
            "$ref": "#/components/parameters/TournamentIdParam"
          },
          {
            "name": "submission_id",
            "in": "query",
            "required": false,
            "description": "Filter runs by specific submission",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "network",
            "in": "query",
            "required": false,
            "description": "Filter runs by test network",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/SkipParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Evaluation runs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvaluationRunResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tournament not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "TournamentIdParam": {
        "name": "tournament_id",
        "in": "path",
        "required": true,
        "description": "Unique tournament identifier",
        "schema": {
          "type": "string",
          "format": "uuid",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        }
      },
      "SkipParam": {
        "name": "skip",
        "in": "query",
        "required": false,
        "description": "Number of records to skip for pagination",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of records to return",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["detail"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "TournamentResponse": {
        "type": "object",
        "required": ["id", "epoch_number", "status", "total_submissions", "total_evaluation_runs", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique tournament identifier"
          },
          "epoch_number": {
            "type": "integer",
            "description": "Epoch number for this tournament"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "running", "completed", "failed"],
            "description": "Current tournament status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the tournament started evaluation"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the tournament completed"
          },
          "total_submissions": {
            "type": "integer",
            "description": "Number of miner submissions"
          },
          "total_evaluation_runs": {
            "type": "integer",
            "description": "Number of evaluation runs executed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the tournament was created"
          }
        }
      },
      "TournamentDetailResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TournamentResponse"
          },
          {
            "type": "object",
            "properties": {
              "config": {
                "type": "object",
                "description": "Tournament configuration parameters",
                "additionalProperties": true
              },
              "test_networks": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "List of blockchain networks used for evaluation"
              },
              "baseline_repository": {
                "type": "string",
                "nullable": true,
                "description": "Git repository URL for baseline implementation"
              },
              "baseline_version": {
                "type": "string",
                "nullable": true,
                "description": "Version/tag of baseline implementation"
              }
            }
          }
        ]
      },
...
      "TournamentResultResponse": {
        "type": "object",
        "description": "Aggregated tournament result for a miner",
        "required": ["id", "tournament_id", "hotkey", "uid", "final_score", "rank"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tournament_id": {
            "type": "string",
            "format": "uuid"
          },
          "hotkey": {
            "type": "string"
          },
          "uid": {
            "type": "integer"
          },
          "output_schema_validity_rate": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true,
            "description": "Percentage of runs passing schema validation"
          },
          "pattern_existence_rate": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true,
            "description": "Percentage of runs finding patterns"
          },
          "feature_performance_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "synthetic_recall_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "pattern_precision_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "novelty_discovery_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "pattern_performance_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "total_runs": {
            "type": "integer",
            "nullable": true
          },
          "total_patterns_reported": {
            "type": "integer",
            "nullable": true
          },
          "total_synthetic_found": {
            "type": "integer",
            "nullable": true
          },
          "total_novelty_valid": {
            "type": "integer",
            "nullable": true
          },
          "total_novelty_invalid": {
            "type": "integer",
            "nullable": true
          },
          "final_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "rank": {
            "type": "integer",
            "minimum": 1
          },
          "beat_baseline": {
            "type": "boolean",
            "description": "Whether this miner beat the baseline"
          },
          "is_winner": {
            "type": "boolean",
            "description": "Whether this is the tournament winner"
          },
          "calculated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LeaderboardResponse": {
        "type": "object",
        "required": ["tournament_id", "epoch_number", "entries"],
        "properties": {
          "tournament_id": {
            "type": "string",
            "format": "uuid"
          },
          "epoch_number": {
            "type": "integer"
          },
          "total_participants": {
            "type": "integer"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentResultResponse"
            }
          }
        }
      },
...
    }
  }
}